+++ src/lib/forms/RadioInline.svelte
@@ -1,9 +1,0 @@
-<div class="flex">
-  <slot />
-</div>
-
-<!--
-  @component
-  [Go to ](https://flowbite-svelte.com/docs/components/)
-  
--->