+++ src/routes/utils/CopyCliboardInput.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
   import Input from '$lib/forms/Input.svelte';
-  import Tooltip from '$lib/tooltips/Tooltip.svelte';
+  import Tooltip from '$lib/tooltip/Tooltip.svelte';
   import Check from './icons/Check.svelte';
   import Clipboard from './icons/Clipboard.svelte';
 
@@ -8,7 +8,7 @@
   const text_copied = 'Copied!';
   const text_not_copied = 'Copy to clipboard';
 
-  let placeholder: string = 'pnpm i flowbite-svelte';
+  let placeholder: string = 'pnpm i -D flowbite-svelte flowbite';
   let tooltip_text: string = text_not_copied;
   let open: boolean | undefined = undefined;
 
@@ -34,7 +34,7 @@
   };
 </script>
 
-<Input size="lg" {placeholder} readonly class="text-sm py-2.5 focus:ring-primary-600 focus:border-primary-600 md:min-w-[255px] {$$props.class ?? ''}">
+<Input size="lg" {placeholder} readonly class="text-sm py-2.5 focus:ring-primary-600 focus:border-primary-600 md:min-w-[315px] {$$props.class ?? ''}">
   <div slot="right" class="flex items-center pl-32">
     <button on:click={copyToClipboard} class="hover:text-primary-700 py-2 px-1">
       {#if tooltip_text == text_not_copied}<Clipboard />{:else}<Check />{/if}